python/shim-service/platform-events_requirements.txt
grpcio
grpcio-tools
protobuf
avro-python3
certifi
